fixes issue where scale is calculated in integers instead of float64

+++ screen/screen.go
@@ -142,7 +142,7 @@ func (s *Screen) Draw(screen *ebiten.Image) {
 
 	// Scaling Screen Image to Render Resolution
 	resX, resY := screen.Bounds().Dx(), screen.Bounds().Dy()
-	scaleX, scaleY := float64(resX/s.ScreenWidth), float64(resY/s.ScreenHeight)
+	scaleX, scaleY := float64(resX)/float64(s.ScreenWidth), float64(resY)/float64(s.ScreenHeight)
 	s.DrawOP.GeoM.Scale(scaleX, scaleY)
 
 	// Render Screen Image to Real Render Screen
